Copilot Edit 1: Refactor styles: Migrate and organize CSS for ItemCard, ItemList, and various pages; remove deprecated styles.
