+++ frontend/src/styles/components/ItemCard.css
@@ -1,0 +1,42 @@
+/* ItemCard standalone styles (subset duplicated from ItemList.css to ensure styling when used alone) */
+
+.item-card {
+  border: none;
+  padding: var(--spacing-lg);
+  border-radius: var(--border-radius-lg);
+  background: var(--white);
+  box-shadow: var(--shadow-sm);
+  transition: all var(--transition-fast);
+  display: flex;
+  flex-direction: row;
+  gap: var(--spacing-md);
+  height: fit-content;
+}
+
+.item-card:hover {
+  background: var(--white);
+  box-shadow: var(--shadow-md);
+  transform: translateY(-2px);
+}
+
+.item-image-container { width: 80px; height: 80px; flex-shrink: 0; }
+.item-thumbnail { width: 100%; height: 100%; object-fit: cover; border-radius: var(--border-radius); border: 1px solid var(--gray-200); }
+.item-content { flex: 1; display: flex; flex-direction: column; }
+
+.item-card h3 { color: var(--primary-color); margin-bottom: var(--spacing-sm); font-size: 1.1rem; font-weight: 600; }
+.item-card p { margin-bottom: var(--spacing-sm); color: var(--gray-600); line-height: 1.5; font-size: 0.9rem; }
+.item-card strong { color: var(--gray-800); margin-right: var(--spacing-xs); }
+
+.item-meta { display: grid; grid-template-columns: 1fr 1fr; gap: var(--spacing-sm); margin-bottom: var(--spacing-md); font-size: 0.85rem; }
+.item-meta-item { display: flex; flex-direction: column; }
+.meta-label { color: var(--gray-500); font-size: 0.75rem; text-transform: uppercase; letter-spacing: 0.5px; margin-bottom: 2px; }
+.meta-value { color: var(--gray-800); font-weight: 600; }
+
+.item-categories { margin-bottom: var(--spacing-md); }
+.category-list { display: flex; flex-wrap: wrap; gap: var(--spacing-xs); margin-top: var(--spacing-xs); }
+.category-item { background: var(--primary-color); color: var(--white); padding: 2px 6px; border-radius: 3px; font-size: 0.7rem; font-weight: 500; display: flex; align-items: center; gap: 4px; }
+.category-icon-small { width: 12px; height: 12px; flex-shrink: 0; filter: brightness(0) invert(1); }
+
+.item-rating { display: flex; align-items: center; gap: var(--spacing-sm); margin-bottom: var(--spacing-md); padding: var(--spacing-sm); background: var(--gray-50); border-radius: var(--border-radius); border: none; }
+.rating-stars { color: #f5b301; font-size: 1.1rem; }
+.rating-info { font-size: 0.85rem; color: var(--gray-600); }
