+++ frontend/src/styles/pages/FrontPage.css
@@ -1,3 +1,44 @@
+/* Page layout scoped to FrontPage only */
+.dashboard-container.front-page {
+  min-height: 100vh;
+  background: var(--gray-100) !important;
+  display: flex;
+  flex-direction: column;
+  font-family: var(--font-primary);
+}
+
+.front-page .dashboard-content {
+  flex: 1;
+  max-width: 1400px;
+  width: 100%;
+  margin: 0 auto;
+  padding: var(--spacing-lg);
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-lg);
+}
+
+.front-page .page-header {
+  text-align: center;
+  background: var(--white);
+  border-radius: var(--border-radius-lg);
+  padding: var(--spacing-xl);
+  box-shadow: var(--shadow-sm);
+  border: 1px solid var(--gray-200);
+  margin-bottom: var(--spacing-lg);
+}
+
+.front-page .page-title {
+  color: var(--gray-800);
+  font-size: 2rem;
+  font-weight: 700;
+  margin: 0 0 var(--spacing-sm) 0;
+  text-align: center;
+}
+
+/* Item management wrapper used by front page */
+.front-page .item-management { width: 100%; max-width: 1200px; margin: 0 auto; }
+.front-page .item-management-layout { display: flex; flex-direction: column; gap: var(--spacing-xl); }
 /* Scope fixes to the front page only */
 .front-page .item-list-section ul {
   display: grid !important;
