+++ frontend/src/styles/pages/FrontPageLayout.css
@@ -1,0 +1,1 @@
+/* Deprecated: merged into FrontPage.css */
