+++ frontend/src/styles/pages/ItemManagement.css
@@ -1,0 +1,33 @@
+/* ItemManagement page-specific styles (layout around NewItemForm and ItemList) */
+
+/* Page layout (migrated from layout/HomePage.css) */
+.dashboard-container { min-height: 100vh; background: var(--gray-100) !important; display: flex; flex-direction: column; font-family: var(--font-primary); }
+.dashboard-content { flex: 1; max-width: 1400px; width: 100%; margin: 0 auto; padding: var(--spacing-lg); display: flex; flex-direction: column; gap: var(--spacing-lg); }
+.page-header { text-align: center; background: var(--white); border-radius: var(--border-radius-lg); padding: var(--spacing-xl); box-shadow: var(--shadow-sm); border: 1px solid var(--gray-200); margin-bottom: var(--spacing-lg); }
+.page-title { color: var(--gray-800); font-size: 2rem; font-weight: 700; margin: 0 0 var(--spacing-sm) 0; text-align: center; }
+
+.item-management {
+  width: 100%;
+  max-width: 1200px;
+  margin: 0 auto;
+}
+
+.item-management-layout {
+  display: flex;
+  flex-direction: column;
+  gap: var(--spacing-xl);
+}
+
+/* Form section container (NewItemForm is responsible for its internal fields) */
+.item-form-section {
+  background: var(--white);
+  border-radius: var(--border-radius-lg);
+  padding: var(--spacing-lg);
+  box-shadow: var(--shadow-sm);
+  border: none;
+  margin-bottom: var(--spacing-lg);
+}
+
+@media (max-width: 480px) {
+  .item-form-section { padding: var(--spacing-md); }
+}
