+++ frontend/src/styles/pages/SearchResults.css
@@ -1,0 +1,179 @@
+/* SearchResults page styles (moved from components/SearchResults.css) */
+
+.search-results-page {
+  padding: 1.5rem;
+  max-width: 800px;
+  margin: 0 auto;
+  font-family: Arial, sans-serif;
+}
+
+.search-results-page h1 {
+  margin-bottom: 3rem;
+  font-size: 1.75rem;
+}
+
+.search-results-page p {
+  margin: 0.75rem 0;
+}
+
+.search-results-page .error {
+  color: #c0392b;
+  font-weight: bold;
+}
+
+/* Search Results List Styles */
+.search-results-list {
+  display: flex;
+  flex-direction: column;
+  gap: 1.5rem;
+  margin-top: 1.5rem;
+}
+
+.search-result-card {
+  display: flex;
+  gap: 1rem;
+  background: white;
+  border-radius: 12px;
+  padding: 1.5rem;
+  box-shadow: 0 2px 8px rgba(0, 0, 0, 0.1);
+  border: 1px solid #e5e7eb;
+  transition: all 0.2s;
+}
+
+.search-result-card:hover {
+  box-shadow: 0 4px 16px rgba(0, 0, 0, 0.15);
+  transform: translateY(-2px);
+}
+
+.search-result-image {
+  flex-shrink: 0;
+  width: 120px;
+  height: 120px;
+}
+
+.result-thumbnail {
+  width: 100%;
+  height: 100%;
+  object-fit: cover;
+  border-radius: 8px;
+  border: 1px solid #e5e7eb;
+}
+
+.search-result-content {
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  gap: 0.75rem;
+}
+
+.result-title-link {
+  text-decoration: none;
+}
+
+.result-title {
+  margin: 0;
+  font-size: 1.25rem;
+  font-weight: 600;
+  color: #1976d2;
+  transition: color 0.2s;
+}
+
+.result-title-link:hover .result-title {
+  color: #1565c0;
+}
+
+.result-description {
+  margin: 0;
+  color: #6b7280;
+  line-height: 1.5;
+  font-size: 0.95rem;
+}
+
+.result-meta {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  flex-wrap: wrap;
+  gap: 1rem;
+}
+
+.result-price {
+  font-size: 1.25rem;
+  font-weight: 700;
+  color: #059669;
+}
+
+.result-rating {
+  display: flex;
+  align-items: center;
+  gap: 0.5rem;
+}
+
+.rating-stars {
+  color: #fbbf24;
+  font-size: 1rem;
+}
+
+.rating-text {
+  color: #6b7280;
+  font-size: 0.875rem;
+}
+
+.result-categories {
+  display: flex;
+  flex-wrap: wrap;
+  gap: 0.5rem;
+}
+
+.result-category-tag {
+  display: flex;
+  align-items: center;
+  gap: 0.25rem;
+  background: #e3f2fd;
+  color: #1976d2;
+  padding: 0.25rem 0.5rem;
+  border-radius: 16px;
+  font-size: 0.75rem;
+  font-weight: 500;
+  border: 1px solid #bbdefb;
+}
+
+.result-category-icon {
+  width: 12px;
+  height: 12px;
+  flex-shrink: 0;
+}
+
+.result-seller {
+  color: #6b7280;
+  font-size: 0.875rem;
+}
+
+.seller-link {
+  color: #1976d2;
+  text-decoration: none;
+  font-weight: 500;
+}
+
+.seller-link:hover {
+  text-decoration: underline;
+}
+
+/* Responsive Design */
+@media (max-width: 640px) {
+  .search-result-card { flex-direction: column; text-align: center; }
+  .search-result-image { align-self: center; width: 100px; height: 100px; }
+  .result-meta { justify-content: center; }
+}
+
+/* Fix for truncated search results */
+.search-results-page .items-container {
+  margin-top: 2rem;
+  max-height: none !important;
+  overflow: visible !important;
+  height: auto !important;
+}
+
+.search-results-page .items-container ul li { margin-bottom: 1.5rem; }
+
+.search-results-page ul { max-height: none !important; overflow: visible !important; }
