+++ frontend/src/styles/pages/UserProfile.css
@@ -1,0 +1,36 @@
+/* UserProfile page styles - reuse layout and list styles to match ItemManagement visuals */
+
+/* Reuse layout containers */
+.dashboard-container { min-height: 100vh; background: var(--gray-100) !important; display: flex; flex-direction: column; font-family: var(--font-primary); }
+.dashboard-content { flex: 1; max-width: 1400px; width: 100%; margin: 0 auto; padding: var(--spacing-lg); display: flex; flex-direction: column; gap: var(--spacing-lg); }
+.page-header { text-align: center; background: var(--white); border-radius: var(--border-radius-lg); padding: var(--spacing-xl); box-shadow: var(--shadow-sm); border: 1px solid var(--gray-200); margin-bottom: var(--spacing-lg); }
+.page-title { color: var(--gray-800); font-size: 2rem; font-weight: 700; margin: 0 0 var(--spacing-sm) 0; text-align: center; }
+
+/* Profile form container */
+.item-form-section { background: var(--white); border-radius: var(--border-radius-lg); padding: var(--spacing-lg); box-shadow: var(--shadow-sm); border: none; margin-bottom: var(--spacing-lg); }
+
+/* Inputs from global.css (form-group, form-input) already present */
+
+/* Follow tabs already styled by FollowTabs.css */
+
+/* Lists and pagination borrowed from layout/HomePage.css */
+.pagination-controls { display: flex; justify-content: center; align-items: center; margin-top: 1.5rem; }
+.pagination-button { all: unset; background-color: var(--gray-100); border-radius: 6px; padding: 0.4rem 0.75rem; margin: 0 0.25rem; font-weight: 600; color: var(--gray-700); cursor: pointer; transition: background-color 0.2s ease; }
+.pagination-button:hover { background-color: var(--gray-300); }
+.pagination-button.active-page { background-color: var(--primary-color); color: white; }
+
+/* Follow tabs (migrated from components/FollowTabs.css) */
+.follow-tabs { display: flex; gap: 1rem; margin: 1.5rem 0; border-bottom: 2px solid #ddd; }
+.follow-tab { background: none; border: none; font-size: 1rem; font-weight: 600; color: var(--gray-600); padding: 0.5rem 0; border-bottom: 3px solid transparent; cursor: pointer; }
+.follow-tab.active { color: var(--gray-900); border-bottom-color: var(--primary-color); }
+.follow-list { list-style: none; padding: 0; margin: 1rem 0; }
+.follow-list li { display: flex; justify-content: space-between; padding: 0.5rem 0; border-bottom: 1px solid #eee; }
+.follow-user-link { color: var(--primary-color); text-decoration: none; font-weight: 500; }
+.unfollow-btn { background: var(--primary-color); color: white; border: none; padding: 0.3rem 0.75rem; border-radius: 5px; cursor: pointer; }
+
+/* Grid container for item cards on profile */
+.item-card-grid { display: grid; gap: 1.5rem; grid-template-columns: repeat(auto-fill, minmax(300px, 1fr)); margin-top: 1rem; }
+
+/* Item management wrappers used on profile page */
+.item-management { width: 100%; max-width: 1200px; margin: 0 auto; }
+.item-management-layout { display: flex; flex-direction: column; gap: var(--spacing-xl); }
